+++ kipher-aes/src/main/kotlin/kipher/aes/internal/AesEncryptionInterface.kt
@@ -3,13 +3,21 @@
 package kipher.aes.internal
 
 internal interface AesEncryptionInterface {
-    fun generateKey(): ByteArray
+    /** Generate a pseudo-random IV. */
     fun generateIv(): ByteArray
 
+    /** Generate a secret key as [ByteArray]. */
+    fun generateKey(): ByteArray
+
+    /** Encrypts the provided [data] using the provided [key]. */
     fun encrypt(data: String, key: ByteArray): ByteArray
+
+    /** Decrypts [encrypted] data using [key]. */
     fun decrypt(encrypted: ByteArray, key: ByteArray): ByteArray
 
-    // With metadata
+    /** Encrypts the provided [data] along with [metadata] using [key]. */
     fun encrypt(data: String, metadata: ByteArray, key: ByteArray): ByteArray
+
+    /** Decrypts [encrypted] data with [metadata] verification using [key]. */
     fun decrypt(encrypted: ByteArray, metadata: ByteArray, key: ByteArray): ByteArray
 }